# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type QueueStatus = typeof QueueStatus[keyof typeof QueueStatus];


export const QueueStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface Queue {
  id: number;
  establishmentName: string;
  status: QueueStatus;
  averageMinutes: number;
  waitingCount: number;
  estimatedMinutes: number;
  createdAt: string;
}

export interface QueueInput {
  /** @minLength 1 */
  establishmentName: string;
  /**
     * @minimum 1
     * @maximum 60
     */
  averageMinutes?: number;
}

export type QueueEntryStatus = typeof QueueEntryStatus[keyof typeof QueueEntryStatus];


export const QueueEntryStatus = {
  waiting: 'waiting',
  called: 'called',
  served: 'served',
  left: 'left',
} as const;

export interface QueueEntry {
  id: number;
  queueId: number;
  name: string;
  position: number;
  peopleAhead: number;
  estimatedMinutes: number;
  joinedAt: string;
  status: QueueEntryStatus;
}

export interface QueueEntryInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
}

export type QueueActivityStatus = typeof QueueActivityStatus[keyof typeof QueueActivityStatus];


export const QueueActivityStatus = {
  waiting: 'waiting',
  called: 'called',
  served: 'served',
  left: 'left',
} as const;

export interface QueueActivity {
  id: number;
  name: string;
  status: QueueActivityStatus;
  timestamp: string;
}

export interface QueueSummary {
  queue: Queue;
  waiting: number;
  called: number;
  servedToday: number;
  averageWaitMinutes: number;
  recentActivity: QueueActivity[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;