# lib/api-zod/src/generated/types/queue.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueStatus } from './queueStatus';

export interface Queue {
  id: number;
  establishmentName: string;
  status: QueueStatus;
  averageMinutes: number;
  waitingCount: number;
  estimatedMinutes: number;
  createdAt: Date;
}