# lib/api-zod/src/generated/types/queueActivity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueActivityStatus } from './queueActivityStatus';

export interface QueueActivity {
  id: number;
  name: string;
  status: QueueActivityStatus;
  timestamp: Date;
}