# lib/api-zod/src/generated/types/queueEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueueEntryStatus } from './queueEntryStatus';

export interface QueueEntry {
  id: number;
  queueId: number;
  name: string;
  position: number;
  peopleAhead: number;
  estimatedMinutes: number;
  joinedAt: Date;
  status: QueueEntryStatus;
}