# lib/api-zod/src/generated/types/queueInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QueueInput {
  /** @minLength 1 */
  establishmentName: string;
  /**
     * @minimum 1
     * @maximum 60
     */
  averageMinutes?: number;
}