# lib/api-zod/src/generated/types/queueSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Queue } from './queue';
import type { QueueActivity } from './queueActivity';

export interface QueueSummary {
  queue: Queue;
  waiting: number;
  called: number;
  servedToday: number;
  averageWaitMinutes: number;
  recentActivity: QueueActivity[];
}